Move i18n and route setups to new convention
