+++ packages/cli/assets/setup/i18n/mock-i18n-types.ts
