+++ packages/cli/assets/setup/routes/locale-check.ts
