+++ packages/cli/src/lib/setups/i18n/replacers.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@shopify/cli-kit/node/fs';
 import {joinPath} from '@shopify/cli-kit/node/path';
 import {ts} from 'ts-morph';
-import {getSkeletonSourceDir} from '../../build.js';
+import {getSetupAssetDir, getSkeletonSourceDir} from '../../build.js';
 import {replaceRemixEnv, replaceServerI18n} from './replacers.js';
 import {DEFAULT_COMPILER_OPTIONS} from '../../transpile/morph/index.js';
 
@@ -40,9 +40,7 @@ describe('i18n replacers', () => {
       await replaceRemixEnv(
         {rootDirectory: tmpDir},
         {},
-        await readFile(
-          fileURLToPath(new URL('./templates/domains.ts', import.meta.url)),
-        ),
+        await readFile(await getSetupAssetDir('i18n', 'domains.ts')),
       );
 
       const newContent = await readFile(joinPath(tmpDir, envDts));
@@ -132,9 +130,7 @@ describe('i18n replacers', () => {
       await replaceServerI18n(
         {rootDirectory: tmpDir, serverEntryPoint: serverTs},
         {},
-        await readFile(
-          fileURLToPath(new URL('./templates/domains.ts', import.meta.url)),
-        ),
+        await readFile(await getSetupAssetDir('i18n', 'domains.ts')),
         false,
       );
