+++ packages/cli/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "remix.env.d.ts"],
+  "include": ["src", "remix.env.d.ts", "assets"],
   "exclude": ["dist", "__tests__", "node_modules", ".turbo"],
   "compilerOptions": {
     "module": "NodeNext",
